fix: return null for all skipped sentinels; extract buildInferentialExtensionsProps helper
Agent-Logs-Url: https://github.com/clarkemoyer/technologyadoptionbarriers.org/sessions/28894258-4a19-4630-88bd-2f9412b74c92
