+++ src/app/results/crp-2026/findings/page.tsx
@@ -12,11 +12,8 @@ import sensitivityData from '@/data/crp-sensitivity-analysis.json'
 import crpValidationData from '@/data/crp-validation.json'
 import EffectSizeChart from '@/components/effect-size-chart'
 import { DATA_UNAVAILABLE } from '@/lib/sentinelMarker'
-import {
-  InferentialExtensions,
-  type InferentialExtensionsProps,
-} from '@/components/results/findings/InferentialExtensions'
-import { findPrimarySample, type ValidationLike } from '@/lib/results/findPrimarySample'
+import { InferentialExtensions } from '@/components/results/findings/InferentialExtensions'
+import { buildInferentialExtensionsProps } from '@/lib/results/buildInferentialExtensionsProps'
 export const metadata: Metadata = {
   title: 'CRP 2026 Key Findings - TABS',
   description:
@@ -122,19 +119,6 @@ const dSize = (d: number | null | undefined): string => {
   if (abs < 0.5) return 'small'
   if (abs < 0.8) return 'medium'
   return 'large'
-}
-
-const buildInferentialExtensionsProps = (validation: unknown): InferentialExtensionsProps => {
-  const sample = findPrimarySample(validation as ValidationLike)
-  if (!sample) return {}
-  return {
-    mediation: sample.mediation_b_r_m as InferentialExtensionsProps['mediation'],
-    perFactorReg: sample.per_factor_regressions as InferentialExtensionsProps['perFactorReg'],
-    standardizedReg:
-      sample.standardized_subfactor_regressions as InferentialExtensionsProps['standardizedReg'],
-    tost: sample.equivalence_test_tost_smb_ent as InferentialExtensionsProps['tost'],
-    powerAnalysis: sample.power_analysis as InferentialExtensionsProps['powerAnalysis'],
-  }
 }
 
 const CrpFindingsPage = () => {
